fix(util.py): non-default branch fails to be checked out
Whenever the --depth option is used, the --single-branch option is implied as well. This means that the default branch is the only branch included in the history in this case. So, if the default branch is "master" then git checkout develop will fail with "pathspec 'develop' did not match any file(s) known to git". Not sure why this didn't fail earlier, but this issue started to appear on the Fine Uploader doc generation server a few days ago. Adding --no-single-branch fixed the issue.